+++ Neat URL-webext/background.js
@@ -37,14 +37,14 @@ let neat_url_hidden_params; // this is an array
 let neat_url_version; // previous version when upgrading, after upgrading the current version
 
 function init(){
-	var valueOrDefault = function(value, defaultValue){
+	let valueOrDefault = function(value, defaultValue){
 		if(value == undefined)
 			return defaultValue;
 		return value;
 	}
 	
-	var valueOrDefaultArray = function(value, defaultValue){
-		var calcValue = valueOrDefault(value, defaultValue);
+	let valueOrDefaultArray = function(value, defaultValue){
+		let calcValue = valueOrDefault(value, defaultValue);
 		if(calcValue == "") return [];
 		return calcValue.split(" ").join("").split(",");
 	}
@@ -70,6 +70,8 @@ function init(){
 
 		browser.browserAction.setBadgeBackgroundColor({color: neat_url_counter_color});
 
+		// Re-initialise listener, hopefully fixes https://github.com/Smile4ever/firefoxaddons/issues/92
+		browser.webRequest.onBeforeRequest.removeListener(cleanURL);
 		if(neat_url_types.length != 0){
 			/// Register for types specified in neat_url_types
 			browser.webRequest.onBeforeRequest.addListener(
@@ -79,7 +81,8 @@ function init(){
 			);
 		}
 
-		initBrowserAction(); // Needs neat_url_icon_theme
+		initBrowserAction(); // needs neat_url_icon_theme
+		initCounter(); // needs neat_url_show_counter
 	}).catch(console.error);
 	
 	browser.storage.local.get([
@@ -178,15 +181,13 @@ function listener(info,tab){
 }
 
 /// Neat URL code
-function upgradeParametersIfNeeded(){	
+function upgradeParametersIfNeeded(){
 	let oldVersion = neat_url_version;
 	let newVersion = browser.runtime.getManifest().version;
 	
 	//console.log("[Neat URL]: upgradeParametersIfNeeded - " + oldVersion + " => " + newVersion);
 	
 	if(oldVersion != newVersion){
-		// Upgrade for neat_url_blocked_params is needed
-		
 		let changes = false;
 		let defaultParams = defaultGlobalBlockedParams.split(", ");
 		
@@ -219,16 +220,11 @@ function upgradeParametersIfNeeded(){
 
 /// Neat URL code
 function removeEndings(leanURL, domain, rootDomain, domainMinusSuffix, blockedParams){
-    let isSearch = leanURL.search == "" ? false : true ;
-    let path = leanURL.pathname;
+	let isSearch = leanURL.search == "" ? false : true ;
+	let path = leanURL.pathname;
 
-	for(let gbp of blockedParams){//Again!? This is third loop!
-		//let firstChar = gbp.substring(0, 1);
-		//if(!isAlpha(firstChar) && firstChar != "_"){//why this now? For $ and $$? Why not check for $ then?
-        if(gbp.startsWith('$')){
-			//let match = getMatch(gbp, domain, rootDomain, domainMinusSuffix, leanURL);//this has already been doone
-			//if(match == "" || match == null) continue;
-
+	for(let gbp of blockedParams){
+		if(gbp.startsWith('$')){
 			path = applyMatch(gbp, isSearch, path);
 		}
 	}
@@ -239,35 +235,22 @@ function removeEndings(leanURL, domain, rootDomain, domainMinusSuffix, blockedPa
 	return leanURL;
 }
 
-// https://stackoverflow.com/questions/2450641/validating-alphabetic-only-string-in-javascript
-/*function isAlpha(str) {//used only once... not even once
-	return /^[a-zA-Z]$/.test(str);//no need for `+` for one char only
-}*/
-
 function applyMatch(match2, isSearch, leanURL){
-	//let firstChar = match2.substr(0, 1);
 	let secondChar = match2.substr(1, 1);
 	let startIndexAsEnd = -1;
 
-	//if(firstChar == "$"){//done earlier
-		if(!isSearch || secondChar == "$" /*|| leanURL.indexOf("/dp/") > -1*/){//what is `/dp/` for???
-			// Check it twice
-			if(match2.indexOf("$") == 0) match2 = match2.substring(1);
-			if(match2.indexOf("$") == 0) match2 = match2.substring(1);
+	if(!isSearch || secondChar == "$"){
+		// Check it twice
+		if(match2.indexOf("$") == 0) match2 = match2.substring(1);
+		if(match2.indexOf("$") == 0) match2 = match2.substring(1);
 
-			// if startIndexAsEnd is -1, we return the original URL without altering it
-			startIndexAsEnd = leanURL.lastIndexOf(match2);
-			//console.log("[Neat URL]: startIndexAsEnd is " + startIndexAsEnd + " inside of " + leanURL + " for " + match2);
+		// if startIndexAsEnd is -1, we return the original URL without altering it
+		startIndexAsEnd = leanURL.lastIndexOf(match2);
+		//console.log("[Neat URL]: startIndexAsEnd is " + startIndexAsEnd + " inside of " + leanURL + " for " + match2);
 
-			if(startIndexAsEnd > -1)
-				leanURL = leanURL.substring(0, startIndexAsEnd);
-		}
-	//}
-
-	/*this has already been doone
-	if(firstChar == "#"){
-		leanURL = leanURL.replace(match2, "");
-	}*/
+		if(startIndexAsEnd > -1)
+			leanURL = leanURL.substring(0, startIndexAsEnd);
+	}
 
 	return leanURL;
 }
@@ -277,9 +260,14 @@ function getMatch(gbp, domain, rootDomain, domainMinusSuffix, detailsUrl){
 		return gbp;
 	}
 
+	// Workaround for https://github.com/Smile4ever/firefoxaddons/issues/76
+	if (gbp == "gws_rd@google.*" && rootDomain == "google.com" && detailsUrl.contains("gws_rd=cr")){
+		return;
+	}
+
 	let keyValue, keyDomain, index;
-    index = gbp.indexOf('@');
-    [keyValue, keyDomain] = [gbp.slice(0,index), gbp.slice(index+1)];
+	index = gbp.indexOf('@');
+	[keyValue, keyDomain] = [gbp.slice(0,index), gbp.slice(index+1)];
 
 	if(keyDomain.startsWith("*.")){
 		// we have a wildcard domain, so compare with root domain please.
@@ -293,6 +281,7 @@ function getMatch(gbp, domain, rootDomain, domainMinusSuffix, detailsUrl){
 	if(keyDomain.endsWith(".*")){
 		//console.log("[Neat URL]: keyDomain " + keyDomain + " ends with .* - domainMinusSuffix is " + domainMinusSuffix);
 		keyDomain = keyDomain.replace(".*", "");
+
 		if ( domainMinusSuffix == keyDomain ) {
 			//console.log("[Neat URL]: matching to wildcard domain");
 			return keyValue;
@@ -309,39 +298,17 @@ function getMatch(gbp, domain, rootDomain, domainMinusSuffix, detailsUrl){
 }
 
 /// Lean URL code
-/*function getParams(URL) { //not used
-    var splitURL = URL.split("?");
-    if ( splitURL.length == 1 ) {
-        return null;
-    }
-
-    var params = {};
-    rawParams = URL.split("?")[1].split("&");
-
-    for ( var i = 0; i < rawParams.length; i++ ) {
-        var rawParam = rawParams[i].split('=');
-        params[rawParam[0]] = rawParam[1];
-    }
-
-    return params;
-}*/
-
-/// Lean URL code
 function buildURL(url, blockedParams, hashParams) {
-    if (blockedParams.length == 0 && hashParams.length == 0) {
-        return url;
-    }
+	if (blockedParams.length == 0 && hashParams.length == 0) {
+		return url;
+	}
 
-	//let urlObject = new URL(url);
-	//let originalSearchParams = urlObject.searchParams;
-
-	/// Process wildcards first for cleaner code
-    //I see no mention about parameter wildcards???. 3.0.0 :/
-	//let wildcardBlockedParams = []; //not used
+	/// Process wildcards parameters
+	// utm_*
 
 	for(let blockedParam of blockedParams){
-        if(blockedParam.startsWith('$')) continue;//another feature -> removeEndings()
-        
+		if(blockedParam.startsWith('$')) continue;//another feature -> removeEndings()
+
 		let wildcardIndex = blockedParam.indexOf("*");
 		if(wildcardIndex == -1) continue;
 
@@ -352,90 +319,50 @@ function buildURL(url, blockedParams, hashParams) {
 		for(let key of url.searchParams.keys()){
 			if(key.startsWith(prefixParam)){
 				// Match! We should remove this parameter.
-                //Here be dragons ;)
-				//console.log("[Neat URL]: buildURL - found wildcard parameter " + pair[0]);
+				// Here be dragons ;)
+				if(neat_url_logging) console.log("[Neat URL]: buildURL - found wildcard parameter " + pair[0]);
 				url.searchParams.delete(key);
 			}
 		}
 	}
 
-	/*for(let wildcardBlockedParam of wildcardBlockedParams){
-		blockedParams.push(wildcardBlockedParam);
-	}*/
-
-	/// Remove blocked url params
-	/*for(let blockedParam of blockedParams){
-		if(blockedParam.indexOf("*") > -1) continue; // Wildcard parameters should have been processed by now
-		originalSearchParams.delete(blockedParam);
-	}*/
-
-	//let newURL = urlObject.href;
-
 	/// Replace hash params
-	
+	const hashUrl = new URL(url.href);
+    hashUrl.search = hashUrl.hash.replace('#', '');
+    hashUrl.hash = '';
+
 	for(let hashParam of hashParams){
-        //this will remove only one, exact, hashParam
-        if(hashParam == url.hash){
-            url.hash = "";
-            break;
+		//this will remove only one, exact, hashParam
+		if(hashParam == "#" + hashUrl.search){
+			hashUrl.search = "";
+			break;
+		}
+
+		/// Do something special!
+		// https://github.com/Smile4ever/firefoxaddons/issues/83
+		// #?utm_source
+		// Example URL: http://www.cuisineactuelle.fr/recettes/mini-burgers-au-foie-gras-331523#utm_source=Facebook&utm_medium=social&utm_campaign=PassionApero
+		if (hashParam.startsWith('#?')) {
+            let specialHashParam = hashParam.replace('#?', '');
+            if(neat_url_logging) console.log("[Neat URL]: buildURL - found hash parameter " + hashParam);
+            hashUrl.searchParams.delete(specialHashParam);
         }
 	}
 
-	/// URL() woes - + is not always %20
-	//this should not matter
-	/*if(decodeURIComponent(url).indexOf("+") == -1){
-		newURL = newURL.replace("+", "%20");
-	}*/
+	const newHash = hashUrl.search.replace('?', '');
+    url.hash = newHash ? `#${newHash}` : '';
 
-	/// URL() woes - Don't encode too much, thank you
-	//???
-	//newURL = urlDecode(newURL);
-
-	/// URL() woes - do not add = when it's not needed
-	/*if(newURL.split("#")[0].replaceAll("=", "") == url.split("#")[0].replaceAll("=", "")){
-		return url;
-	}*/
-
-	/// URL() woes - Don't encode too much, thank you
-	/*if(newURL.indexOf("??") == -1 && url.indexOf("??") > -1){
-		return url;
-	}*/
-
-    return url;
+	return url;
 }
 
-/// Neat URL code
-// Copied from https://stackoverflow.com/questions/8498592/extract-hostname-name-from-string
-/*function getDomain(url) { //not used
-	if(url == undefined || url == null) return null;
-	
-    var hostname = url.replace("www.", ""); // leave www out of this discussion. I don't consider this a subdomain
-    //find & remove protocol (http, ftp, etc.) and get hostname
-    
-    if (url.indexOf("://") > -1) {
-        hostname = hostname.split('/')[2];
-    }
-    else {
-        hostname = hostname.split('/')[0];
-    }
-
-    //find & remove port number
-    hostname = hostname.split(':')[0];
-    //find & remove "?"
-    hostname = hostname.split('?')[0];
-
-    return hostname;
-}*/
-
 function getDomainMinusSuffix(domain){
-	//let domain = getDomain(url);
 	let lastIndex = domain.lastIndexOf(".");
 	let previousLastIndex = domain.lastIndexOf(".", lastIndex - 1);
   
 	if(lastIndex - previousLastIndex < 4){
 		lastIndex = previousLastIndex;
 	}
-	
+
 	return domain.substring(0, lastIndex);
 }
 
@@ -444,65 +371,51 @@ function getDomainMinusSuffix(domain){
 function getRootDomain(domain) {
 	if(domain == undefined || domain == null) return null;
 
-    let /*domain = getDomain(url),*/
-        splitArr = domain.split('.'),
-        arrLen = splitArr.length;
+	let /*domain = getDomain(url),*/
+		splitArr = domain.split('.'),
+		arrLen = splitArr.length;
 
-    // Extract the root domain
-    //http://publicsuffix.org/list/
-    if (arrLen > 2) {
-        if(splitArr[arrLen - 2].length <= 3){
-            // oops.. this is becoming an invalid URL
-            // Example URLs that trigger this code path are https://images.google.co.uk and https://google.co.uk
-            domain = splitArr[arrLen - 3] + '.' + splitArr[arrLen - 2] + '.' + splitArr[arrLen - 1];
-        }else{
-            domain = splitArr[arrLen - 2] + '.' + splitArr[arrLen - 1];
-        }
-    }
-    
-    return domain;
+	// Extract the root domain
+	//http://publicsuffix.org/list/
+	if (arrLen > 2) {
+		if(splitArr[arrLen - 2].length <= 3){
+			// oops.. this is becoming an invalid URL
+			// Example URLs that trigger this code path are https://images.google.co.uk and https://google.co.uk
+			domain = splitArr[arrLen - 3] + '.' + splitArr[arrLen - 2] + '.' + splitArr[arrLen - 1];
+		}else{
+			domain = splitArr[arrLen - 2] + '.' + splitArr[arrLen - 1];
+		}
+	}
+
+	return domain;
 }
-
-/*function urlDecode(url){ //not used
-	return url
-		.replaceAll("%3d", "=").replaceAll("%3D", "=")
-		.replaceAll("%2c", ",").replaceAll("%2C", ",")
-		.replaceAll("%3A", ":").replaceAll("%3A", ":")
-		.replaceAll("%2f", "/").replaceAll("%2F", "/");
-}*/
 
 /// Lean URL / Neat URL code
 function cleanURL(details) {
 	if(!enabled) return;
 
-	let originalDetailsUrl = details.url;
 	let url = new URL(details.url);
-    let domain = url.hostname
+	let domain = url.hostname;
+	let originalDetailsUrl = details.url;
+
 	// Do not change links for these domains
 	for(let blackDomain of neat_url_blacklist){
-		//console.log("[Neat URL]: blackDomain " + blackDomain);
 		if(domain.endsWith(blackDomain)){
-			if(neat_url_logging){
-				console.log(`[Neat URL]: not rewriting '${url.href}'`);
-			}
+			if(neat_url_logging) console.log(`[Neat URL]: not rewriting '${url.href}'`);
 			return;
 		}
 	}
 
-    //let baseURL = url.split("?")[0];
-    //let params = url.search;
-    //let hashParams = url.hash;
-    
 	if ("" === url.search && "" === url.hash){
-		//console.log(`[Neat URL]: no params for '${url.href}'`);
+		if(neat_url_logging) console.log(`[Neat URL]: no params for '${url.href}'`);
 		return;
 	}
 	//console.log(`[Neat URL]: processing '${url.href}'`);
 
-	//var domain = getDomain(url);
+	//let domain = getDomain(url);
 	domain = domain.replace(/^www\./i, '');//getDomain() -> //leave www out of this discussion. I don't consider this a subdomain
-	var rootDomain = getRootDomain(domain);
-	var domainMinusSuffix = getDomainMinusSuffix(domain);
+	let rootDomain = getRootDomain(domain);
+	let domainMinusSuffix = getDomainMinusSuffix(domain);
 
 	if (domain == null || rootDomain == null || domainMinusSuffix == null ){
 		return;
@@ -516,6 +429,7 @@ function cleanURL(details) {
 		if(match == "" || match == null) continue;
 
 		if(match.startsWith("#")){
+			if(neat_url_logging) console.log(`[Neat URL]: hash param '${match}' matches`);
 			hashParams.push(match);
 			continue;
 		}
@@ -532,47 +446,30 @@ function cleanURL(details) {
 		//console.log(`[Neat URL]: if includes(${key})`)
 		if (blockedParams.includes(key)) {
 			//console.log(`[Neat URL]: delete(${key})`)
-            url.searchParams.delete(key);
+			url.searchParams.delete(key);
 		}
 	}
 
 	// https://github.com/Smile4ever/firefoxaddons/issues/30 should no longer occur with the new buildURL function
 	// https://github.com/Smile4ever/firefoxaddons/issues/47 should be solved as well
-    leanURL = buildURL(url, blockedParams, hashParams);
-    leanURL = removeEndings(leanURL, domain, rootDomain, domainMinusSuffix, blockedParams);
+	leanURL = buildURL(url, blockedParams, hashParams);
+	leanURL = removeEndings(leanURL, domain, rootDomain, domainMinusSuffix, blockedParams);
 
 	// Is the URL changed?
-	/*let decodedDetailsURL = urlDecode(decodeURIComponent(url));
-	let decodedOriginalDetailsURL = urlDecode(decodeURIComponent(originalDetailsUrl));
-	let decodedLeanURL = urlDecode(decodeURIComponent(leanURL));
-
-	if(decodedLeanURL == decodedDetailsURL || decodedLeanURL == decodedOriginalDetailsURL){
-		return;
-	}*/
-    if(originalDetailsUrl == leanURL.href) return;
-
-	// Don't change the URL if any of these is true
-	/*if(decodedLeanURL.indexOf("utm.gif") > -1 || decodedDetailsURL.indexOf("&&") > -1){//will see
-		return;
-	}*/
+	if(originalDetailsUrl == leanURL.href) return;
 
 	if(neat_url_logging){
 		console.log(`[Neat URL]: (type ${details.type}): '${originalDetailsUrl}' has been changed to '${leanURL}'`);
 	}
 
-	//let leanURLDomain = leanURL.hostname;
-
 	const applyAfter = 1000;
-    // Animate the toolbar icon
-    if(leanURL.hostname != "addons.mozilla.org"){
+
+	if(leanURL.hostname != "addons.mozilla.org"){
 		animateToolbarIcon();
-		incrementBadgeValue(details.tabId);
+		if(neat_url_show_counter) incrementBadgeValue(details.tabId);
 	} else {
-
-    // webRequest blocking is not supported on mozilla.org, lets fix this
-    // but only if we are navigating to addons.mozilla.org and there doesn't exist a tab yet with the same URL
-
-	//if(leanURLDomain == "addons.mozilla.org"){
+		// webRequest blocking is not supported on mozilla.org, lets fix this
+		// but only if we are navigating to addons.mozilla.org and there doesn't exist a tab yet with the same URL
 		if(details.type != "main_frame") return;
 		if(globalNeatURL == leanURL.href) return;
 
@@ -595,7 +492,7 @@ function cleanURL(details) {
 						}
 						browser.tabs.update(tab.id, {url: globalNeatURL});//May be fired more than once?
 						animateToolbarIcon();
-						incrementBadgeValue(globalTabId);
+						if(neat_url_show_counter) incrementBadgeValue(globalTabId);
 					}
 				}
 
@@ -607,54 +504,7 @@ function cleanURL(details) {
 		}, applyAfter);
 	}
 
-    return { redirectUrl: leanURL.href };
-}
-
-/// Neat URL code
-function animateToolbarIcon(){
-	if(neat_url_icon_animation == "none") return;
-	
-	var defaultState = resolveIconURL("neaturl-96-state0.png");
-	var images = [];
-	var imagesMissingUnderscore = [resolveIconURL("neaturl-96-state-1.png"), defaultState, resolveIconURL("neaturl-96-state-1.png")];
-	var imagesRotate = [resolveIconURL("neaturl-96-state1.png"), resolveIconURL("neaturl-96-state2.png"), resolveIconURL("neaturl-96-state3.png")];
-
-	if(neat_url_icon_animation == "missing_underscore")
-		images = imagesMissingUnderscore;
-	if(neat_url_icon_animation == "rotate")
-		images = imagesRotate;
-		
-	if(images.length == 0 && neat_url_icon_animation == "surprise_me"){
-		// https://stackoverflow.com/questions/4959975/generate-random-number-between-two-numbers-in-javascript
-		var zeroOrOne = Math.floor(Math.random() * 2);
-		
-		if(zeroOrOne == 0)
-			images = imagesMissingUnderscore;
-		if(zeroOrOne == 1)
-			images = imagesRotate;
-	}
-	
-	var time = 350 * images.length;
-	
-	// Set first state
-	browser.browserAction.setIcon({path: images[0]});
-		
-	if(images.length > 1){
-		setTimeout(function(){
-			browser.browserAction.setIcon({path: images[1]});
-		}, 1 * time);
-	}
-
-	if(images.length > 2){
-		setTimeout(function(){
-			browser.browserAction.setIcon({path: images[2]});
-		}, 2 * time);
-	}
-	
-	// Reset to default state
-	setTimeout(function(){
-		browser.browserAction.setIcon({path: defaultState});
-	}, 3 * time);
+	return { redirectUrl: leanURL.href };
 }
 
 /// Translate Now code
@@ -668,97 +518,7 @@ function notify(message){
 	});
 }
 
-function resolveIconByTheme(file, theme){
-	return "icons/" + theme + "/" + file;
-}
-
-function resolveIconURL(file){
-	let theme = neat_url_icon_theme.replace("_notiflight", "").replace("_notifdark", "");
-	return resolveIconByTheme(file, theme);
-}
-
-function resolveIconUrlNotif(file){
-	if(neat_url_icon_theme.indexOf("_notiflight") > -1){
-		return resolveIconByTheme(file, "light");
-	}
-	if(neat_url_icon_theme.indexOf("_notifdark") > -1){
-		return resolveIconByTheme(file, "dark");
-	}
-	return resolveIconByTheme(file, neat_url_icon_theme);
-}
-
 String.prototype.replaceAll = function(search, replacement) {
-	var target = this;
+	let target = this;
 	return target.replace(new RegExp(search, 'g'), replacement);
 };
-
-function updateBadgeText(tabId){
-	// We're only updating the view
-	// The data won't be touched
-
-	//console.log("[Neat URL]: updateBadgeText - I'm updating the badge text for tabId " + tabId);
-	let badgeCounts = badge[tabId];
-	if(badgeCounts == null){
-		//console.log("[Neat URL]: badgeCounts == null");
-		badgeCounts = ""; // Set empty instead of null or 0
-	}else{
-		//console.log("[Neat URL]: badgeCounts is not null");
-	}
-
-	/// Update browserAction with badge count for the current tab. If the current tab changes, we will update it again
-	//console.log("[Neat URL]: updateBadgeText - badgeCounts is " + badgeCounts);
-	browser.browserAction.setBadgeText({text: badgeCounts+""});
-}
-
-function incrementBadgeValue(tabId){
-	// We're only updating the data
-	// The view will be updated when needed by the listeners
-
-	if(tabId == -1)
-	{
-		// The request is unrelated to a tab
-		return;
-	}
-
-	//console.log("[Neat URL]: getting badgeCounts for tabId " + tabId + " is " + badge[tabId]);
-	let badgeCounts = badge[tabId];
-	if(badgeCounts == null){
-		//console.log("[Neat URL]: badgeCounts == null");
-		badgeCounts = 0;
-	}else{
-		//console.log("[Neat URL]: badgeCounts is not null, but " + badgeCounts);
-	}
-	badgeCounts++;
-	badge[tabId] = badgeCounts;
-
-	//console.log("[Neat URL]: setting badgeCount to " + badgeCounts + " - result is " + badge[tabId]);
-}
-
-browser.tabs.onCreated.addListener(function(tab){
-	// Tabs can be created in the background. We're not interested in that.
-	if(tab.active){
-		//console.log("[Neat URL]: Updating currentTabId to " + tab.id + " by onCreated");
-		updateBadgeText(tab.id);
-	}
-});
-
-browser.tabs.onUpdated.addListener(function(tabId, changeInfo, tabInfo){
-	// Tabs can be updated in the background. We're not interested in that.
-	if(tabInfo.active){
-		//console.log("[Neat URL]: Updating currentTabId to " + tabInfo.id + " by onUpdated");
-		updateBadgeText(tabInfo.id);
-	}
-});
-
-browser.tabs.onActivated.addListener(function(activeInfo){
-	// Tabs can be inactive when updating or creating and afterwards activated. We could check tab.active but it will always be true
-	//console.log("[Neat URL]: Updating currentTabId to " + activeInfo.tabId + " by onActivated");
-	updateBadgeText(activeInfo.tabId);
-});
-
-// Reset inside "dictionary"
-browser.tabs.onRemoved.addListener(function(tab){
-	// We're not changing the currentTabId here, since onRemoved will call onActivated implicitly
-	// We're only resetting the count for the badge, so that if another tab gets the same id, we don't be using the badge count of that unrelated tab
-	badge[tab.id] = null;
-});